Read the core body.
In Java the core body is pre-built by a
class (com.argot.meta.MetaLoader) instead of build built up through
definitions in the dictionary. There is a check that the core (in
memory) and the core (read from dictionary) are equal, this is achieved
by writing the current core to a byte array and running a comparison.

I doubt it's possible to bootstrap all of the Meta defintions, but maybe
we can bootstrap some of them instead of having them predefined and
fixed. Hmm.