Record and use readers to read the dictionaries
It is clear now that the way the dictionaries are formed refers only to
the known meta model (version 1.3) and that the stream ids used during
reading only refer to that.

At the current stage the stream id -> definition name mapping is global,
and this leads to the later stream ids (definitely final, maybe messages
as well) overwriting the library mappings. But the mapping from binary
to library still requires the core mappings in place. These will have to
be seperated out into a 'reading a dictionary' type mapping library and
an everything else library. I guess this relates to what the Java app
calls the "reference_map".

Until then, we still have the custom fallbacks in dictionary.js