Add prep work and notes for building structures
No functional differences here, just a few comments on what we'll have
to do. The outcome is, on registration of a new item in the library,
record against the item whether it iis a name, definition or relation.

When we build structures we should build them in name > defintion >
relation order.