Prep for building our own readers and writers
It appears the meta reference is the element that indicates something is
made up of subcomponents. After building a structure we will need to
descend into it and build the reader. The descent may be arbitrarily
deep and that could make difficulties, perhaps the meta references are
only contained in one other type (or a small subset of other types), and
if that's the case we may be able to build readers (and writers) inside
the dictionary readers. Hmm

As a sidenote, this descent would be the perfect case for statically
type pattern matching. Woe.js