Build data reader functions
Currently this is implemented against the light dictionary example,
reading a data structure taken from the Java tests ([120, 10, 30]) and
ensuring that the object created contains the correct colors.

The test data reader loops over an array of readers, one for each color,
with the result from each read being added to an object using the
description name. Thus 120, 10, 30 becomes {red:120,green:10,blue:30}.

There will no doubt be other structures that need to be read, some
investigation (or alternative test dictionary) will be required before
those can be accomodated. And, of course, we still need to implement the
writers.