Add test for the set_colour nested data structure
The set_colour data structure is a simple wrap arund the colour
structure and has exactly the same binary format, it is only the intent
of the reader or writer that determines what is output.

This test shows an example of nested structures (set_colour containing
colour containing an array of uint8s) and that it can be read as
expected.

There is still a conflict around the reader functions that is resulting
in the two different properties on the type definition (reader and
typeReader) and, as such, there's a required workaround in the test to
map the colour definition typeReader onto a reader property.