Remove log warnings on already known definitions
The default starting library already contains definitions for a few
dictionary types that are required to bootstrap reading of the
dictionaries.

These definitions are still contained in the dictionary files, so when
we hit them we can build an actual representation and throw away those
that we fabricated. We don't want to warn on this occasion.

The check that the existing name is not equal to the new name helps us
here, but perhaps we should be more specific and hardcode these default
definitions.

In any case, if a dictionary tries to override a definition with a new
name then we will experience a warning.