Stop calling callbacks with promises
The load dictionary and read message functions were both accepting
callbacks and then calling them with promise values, an unessessary
duplication of effort and source of confusion.

This change removes the callbacks and has the function each just return
a promise. I was unsure if this route was acceptable previously because
the nessessary path to accomplish this appeared to result with promises
nested inside promises but, fortunately, q handles all of this with ease
and does not nest.