Throw error trying to load file in not-node env
When we run in the browser with browserify we can't use the fs module,
so we need a way to accomodate that. For now we can throw an error
saying that it isn't handled instead of trying to go ahead and call an
fs.function that doesn't exist.

It may make more sense to create two argot projects in npm, with the
second one explicity being for the browser. But until then we'll see
where this route takes us.